shacl_validation::shacl_engine::engine

Trait Engine

pub trait Engine<S: NeighsRDF> {
    // Required methods
    fn evaluate(
        &mut self,
        store: &S,
        shape: &ShapeIR,
        component: &ComponentIR,
        value_nodes: &ValueNodes<S>,
        source_shape: Option<&ShapeIR>,
        maybe_path: Option<SHACLPath>,
        shapes_graph: &SchemaIR,
    ) -> Result<Vec<ValidationResult>, Box<ValidateError>>;
    fn target_node(
        &self,
        store: &S,
        node: &Object,
    ) -> Result<FocusNodes<S>, Box<ValidateError>>;
    fn target_class(
        &self,
        store: &S,
        class: &Object,
    ) -> Result<FocusNodes<S>, Box<ValidateError>>;
    fn target_subject_of(
        &self,
        store: &S,
        predicate: &IriS,
    ) -> Result<FocusNodes<S>, Box<ValidateError>>;
    fn target_object_of(
        &self,
        store: &S,
        predicate: &IriS,
    ) -> Result<FocusNodes<S>, Box<ValidateError>>;
    fn implicit_target_class(
        &self,
        store: &S,
        shape: &Object,
    ) -> Result<FocusNodes<S>, Box<ValidateError>>;
    fn record_validation(
        &mut self,
        node: Object,
        shape_idx: ShapeLabelIdx,
        results: Vec<ValidationResult>,
    );
    fn has_validated(&self, node: &Object, shape_idx: ShapeLabelIdx) -> bool;
    fn get_cached_results(
        &self,
        node: &Object,
        shape_idx: ShapeLabelIdx,
    ) -> Option<&Vec<ValidationResult>>;

    // Provided methods
    fn build_indexes(&mut self, _store: &S) -> Result<(), Box<ValidateError>> { ... }
    fn focus_nodes(
        &self,
        store: &S,
        targets: &[CompiledTarget],
    ) -> Result<FocusNodes<S>, Box<ValidateError>> { ... }
    fn path(
        &self,
        store: &S,
        shape: &PropertyShapeIR,
        focus_node: &S::Term,
    ) -> Result<FocusNodes<S>, Box<ValidateError>> { ... }
}

Required Methods

fn evaluate( &mut self, store: &S, shape: &ShapeIR, component: &ComponentIR, value_nodes: &ValueNodes<S>, source_shape: Option<&ShapeIR>, maybe_path: Option<SHACLPath>, shapes_graph: &SchemaIR, ) -> Result<Vec<ValidationResult>, Box<ValidateError>>

fn target_node( &self, store: &S, node: &Object, ) -> Result<FocusNodes<S>, Box<ValidateError>>

If s is a shape in a shapes graph SG and s has value t for sh:targetNode in SG then { t } is a target from any data graph for s in SG.

fn target_class( &self, store: &S, class: &Object, ) -> Result<FocusNodes<S>, Box<ValidateError>>

fn target_subject_of( &self, store: &S, predicate: &IriS, ) -> Result<FocusNodes<S>, Box<ValidateError>>

fn target_object_of( &self, store: &S, predicate: &IriS, ) -> Result<FocusNodes<S>, Box<ValidateError>>

fn implicit_target_class( &self, store: &S, shape: &Object, ) -> Result<FocusNodes<S>, Box<ValidateError>>

fn record_validation( &mut self, node: Object, shape_idx: ShapeLabelIdx, results: Vec<ValidationResult>, )

fn has_validated(&self, node: &Object, shape_idx: ShapeLabelIdx) -> bool

fn get_cached_results( &self, node: &Object, shape_idx: ShapeLabelIdx, ) -> Option<&Vec<ValidationResult>>

Returns the cached validation results for a given (node, shape_idx) pair, if any.

Provided Methods

fn build_indexes(&mut self, _store: &S) -> Result<(), Box<ValidateError>>

Pre-builds internal indexes from the data graph for faster target resolution.

This should be called once before the validation loop starts.

fn focus_nodes( &self, store: &S, targets: &[CompiledTarget], ) -> Result<FocusNodes<S>, Box<ValidateError>>

fn path( &self, store: &S, shape: &PropertyShapeIR, focus_node: &S::Term, ) -> Result<FocusNodes<S>, Box<ValidateError>>